# openapi-dump.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/user/getUserList": {
      "get": {
        "operationId": "UserController_getUserList",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "页码",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "description": "每页条数",
            "schema": {
              "minimum": 1,
              "default": 10,
              "example": 10,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseWrapperDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "allOf": [
                            {
                              "$ref": "#/components/schemas/PaginatedDataDto"
                            },
                            {
                              "properties": {
                                "list": {
                                  "type": "array",
                                  "items": {
                                    "$ref": "#/components/schemas/UserListItemDto"
                                  }
                                }
                              }
                            }
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误或业务校验失败"
          },
          "401": {
            "description": "未登录"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "summary": "获取用户列表",
        "tags": [
          "api/user"
        ]
      }
    },
    "/api/user/getUserById": {
      "get": {
        "operationId": "UserController_getUserById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "query",
            "description": "用户ID",
            "schema": {
              "example": 1,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseWrapperDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserDetailDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误或业务校验失败"
          },
          "401": {
            "description": "未登录"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "summary": "获取用户详情（含角色）",
        "tags": [
          "api/user"
        ]
      }
    },
    "/api/user/addUser": {
      "post": {
        "operationId": "UserController_addUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseWrapperDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserBriefDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误或业务校验失败"
          },
          "401": {
            "description": "未登录"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "summary": "新增用户",
        "tags": [
          "api/user"
        ]
      }
    },
    "/api/user/updateUser": {
      "post": {
        "operationId": "UserController_updateUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/UpdateUserDto"
                  },
                  {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "number",
                        "description": "用户ID",
                        "example": 1
                      }
                    },
                    "required": [
                      "id"
                    ]
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseWrapperDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserBriefDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误或业务校验失败"
          },
          "401": {
            "description": "未登录"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "summary": "更新用户",
        "tags": [
          "api/user"
        ]
      }
    },
    "/api/user/delUser": {
      "post": {
        "operationId": "UserController_delUser",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "description": "用户ID",
                    "example": 1
                  }
                },
                "required": [
                  "id"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseWrapperDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "message": {
                              "type": "string",
                              "description": "操作结果消息"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误或业务校验失败"
          },
          "401": {
            "description": "未登录"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "summary": "删除用户（软删除）",
        "tags": [
          "api/user"
        ]
      }
    },
    "/api/user/assignRoles": {
      "post": {
        "operationId": "UserController_assignRoles",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "number",
                    "description": "用户ID",
                    "example": 1
                  },
                  "roleIds": {
                    "type": "array",
                    "items": {
                      "type": "number"
                    },
                    "description": "角色ID数组",
                    "example": [
                      1,
                      2
                    ]
                  }
                },
                "required": [
                  "userId",
                  "roleIds"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponseWrapperDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "message": {
                              "type": "string",
                              "description": "操作结果消息"
                            }
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误或业务校验失败"
          },
          "401": {
            "description": "未登录"
          },
          "403": {
            "description": "权限不足"
          }
        },
        "summary": "为用户分配角色",
        "tags": [
          "api/user"
        ]
      }
    },
...
  },
  "info": {
    "title": "后台管理服务",
    "description": "",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "connect.sid"
      }
    },
    "schemas": {
      "ApiResponseWrapperDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "number",
            "description": "业务状态码，0=成功",
            "example": 0
          },
          "message": {
            "type": "string",
            "description": "提示消息",
            "example": "操作成功"
          },
          "success": {
            "type": "boolean",
            "description": "是否成功",
            "example": true
          },
          "data": {
            "type": "object",
            "description": "业务数据"
          }
        },
        "required": [
          "code",
          "message",
          "success",
          "data"
        ]
      },
...
      "PaginatedDataDto": {
        "type": "object",
        "properties": {
          "list": {
            "description": "数据列表",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "total": {
            "type": "number",
            "description": "总条数",
            "example": 100
          },
          "page": {
            "type": "number",
            "description": "当前页码",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "description": "每页条数",
            "example": 10
          }
        },
        "required": [
          "list",
          "total",
          "page",
          "pageSize"
        ]
      },
      "UserListItemDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "用户ID",
            "example": 1
          },
          "username": {
            "type": "string",
            "description": "用户名",
            "example": "zhangsan"
          },
          "email": {
            "type": "string",
            "description": "邮箱",
            "example": "test@example.com"
          },
          "phone": {
            "type": "string",
            "description": "手机号",
            "example": "13800138000"
          },
          "status": {
            "type": "number",
            "description": "状态",
            "enum": [
              0,
              1
            ],
            "example": 1
          },
          "createdAt": {
            "type": "string",
            "description": "创建时间",
            "example": "2026-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "description": "更新时间",
            "example": "2026-06-24T12:00:00.000Z"
          }
        },
        "required": [
          "id",
          "username",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "RoleDataDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "角色ID",
            "example": 1
          },
          "name": {
            "type": "string",
            "description": "角色名称",
            "example": "超级管理员"
          },
          "code": {
            "type": "string",
            "description": "角色编码",
            "example": "super_admin"
          },
          "description": {
            "type": "string",
            "description": "角色描述",
            "example": "拥有所有权限"
          },
          "status": {
            "type": "number",
            "description": "状态",
            "enum": [
              0,
              1
            ],
            "example": 1
          },
          "createdAt": {
            "type": "string",
            "description": "创建时间",
            "example": "2026-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "description": "更新时间",
            "example": "2026-06-24T12:00:00.000Z"
          }
        },
        "required": [
          "id",
          "name",
          "code",
          "status",
          "createdAt",
          "updatedAt"
        ]
      },
      "UserRoleJoinDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number",
            "description": "用户ID",
            "example": 1
          },
          "roleId": {
            "type": "number",
            "description": "角色ID",
            "example": 2
          },
          "role": {
            "description": "角色详情",
            "allOf": [
              {
                "$ref": "#/components/schemas/RoleDataDto"
              }
            ]
          }
        },
        "required": [
          "userId",
          "roleId",
          "role"
        ]
      },
      "UserDetailDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "用户ID",
            "example": 1
          },
          "username": {
            "type": "string",
            "description": "用户名",
            "example": "zhangsan"
          },
          "email": {
            "type": "string",
            "description": "邮箱",
            "example": "test@example.com"
          },
          "phone": {
            "type": "string",
            "description": "手机号",
            "example": "13800138000"
          },
          "status": {
            "type": "number",
            "description": "状态",
            "enum": [
              0,
              1
            ],
            "example": 1
          },
          "createdAt": {
            "type": "string",
            "description": "创建时间",
            "example": "2026-01-01T00:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "description": "更新时间",
            "example": "2026-06-24T12:00:00.000Z"
          },
          "userRoles": {
            "description": "关联角色列表（含连接表字段）",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserRoleJoinDto"
            }
          }
        },
        "required": [
          "id",
          "username",
          "status",
          "createdAt",
          "updatedAt",
          "userRoles"
        ]
      },
      "UserBriefDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "description": "用户ID",
            "example": 1
          },
          "username": {
            "type": "string",
            "description": "用户名",
            "example": "zhangsan"
          },
          "email": {
            "type": "string",
            "description": "邮箱",
            "example": "test@example.com"
          },
          "phone": {
            "type": "string",
            "description": "手机号",
            "example": "13800138000"
          },
          "status": {
            "type": "number",
            "description": "状态（addUser 不返回此字段）",
            "enum": [
              0,
              1
            ],
            "example": 1
          }
        },
        "required": [
          "id",
          "username"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "用户名",
            "example": "zhangsan"
          },
          "password": {
            "type": "string",
            "description": "密码（明文，服务端 bcrypt 哈希后入库）",
            "example": "Abc12345"
          },
          "email": {
            "type": "string",
            "description": "邮箱",
            "example": "test@example.com"
          },
          "phone": {
            "type": "string",
            "description": "手机号",
            "example": "13800138000"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "用户名",
            "example": "zhangsan"
          },
          "password": {
            "type": "string",
            "description": "新密码（非空时触发密码重哈希）",
            "example": "NewPass123"
          },
          "email": {
            "type": "string",
            "description": "邮箱",
            "example": "newemail@example.com"
          },
          "phone": {
            "type": "string",
            "description": "手机号",
            "example": "13900139000"
          },
          "status": {
            "type": "number",
            "description": "用户状态",
            "enum": [
              0,
              1
            ],
            "example": 1
          }
        }
      },
...
    }
  }
}